Web Registration with User Self-Help Panel( NOT finished )
This is modified by cym0301. Web Registration page is set. User
Self-Help Panel is still under construction with a page showing UserInfo
finished. All are not for business use. Feel free to use it or modify it
to make it better. Kindly sharing the modification and making a pull
request are welcome and appreciated.